backend1/requirements.txt
Flask
flask-cors
mongoengine
python-dotenv
